rp5_service/requirements.txt
requests
psutil
time
json
subprocess
python-dotenv

